Move to Config::Onion-based configuration.
While we are not using the full power of Config::Onion, this gives us
room to grow this out later if necessary.

Also allowes us to drop the share directory and reading therefrom
without impact on our ability to properly test the module or requiring
contortions to make it work.

All of this leads to renaming the config file verification function to
better express what it is actually doing (verifying the config instead
of the file.)